+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add tests for job categories import fix
 - Subheading block (BE)
 
 ## 14.0 (2020-04-22)